add separeted service to file compression
