+++ src/file-compression/file-compression.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { FileCompressionService } from './file-compression.service.js';
+
+@Module({
+  providers: [FileCompressionService],
+  exports: [FileCompressionService],
+})
+export class FileProcessingModule {}
